requirements.txt
asgiref==3.2.3
autopep8==1.4.4
Django==3.0
django-environ==0.4.5
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
mccabe==0.6.1
model-mommy==2.0.0
Pillow==7.0.0
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.3
sqlparse==0.3.0
